fix: 폴더링 컨벤션대로 수정 (#9)
