+++ src/main/java/com/acon/server/review/application/mapper/ReviewImageMapper.java
@@ -1,7 +1,7 @@
 package com.acon.server.review.application.mapper;
 
-import com.acon.server.member.infra.entity.review.ReviewImageEntity;
 import com.acon.server.review.domain.entity.ReviewImage;
+import com.acon.server.review.infra.entity.ReviewImageEntity;
 import org.mapstruct.Mapper;
 import org.mapstruct.ReportingPolicy;
